+++ src/common/js/configure.js
@@ -2,11 +2,11 @@
  * pwix:modal/src/common/js/configure.js
  */
 
-_defaults = {
+pwixModal._defaults = {
     verbosity: MD_VERBOSE_NONE
 };
 
-pwixModal._conf = { ...defaults };
+pwixModal._conf = { ...pwixModal._defaults };
 
 /**
  * @summary Package configuration
@@ -15,7 +15,7 @@ pwixModal._conf = { ...defaults };
  *  Should be *in same terms* called both by the client and the server.
  */
 pwixModal.configure = function( o ){
-    pwixModal._conf = { ...defaults, ...o };
+    pwixModal._conf = { ...pwixModal._defaults, ...o };
 
     if( pwixModal._conf.verbosity & MD_VERBOSE_CONFIGURE ){
         console.log( 'pwix:modal configure() with', o, 'building', pwixModal._conf );